==> startup/3 pipeline.ipynb <==
import pandas as pd

df = pd.read_csv('processed.csv')
df

df2 = df.drop(['Unnamed: 0','state_code','zip_code','city','name','founded_at','last_funding_at','age_first_funding_year','is_CA','is_NY','is_MA','is_TX','category_code','is_software','is_web','is_mobile',
                'is_enterprise','is_advertising','is_gamesvideo','is_ecommerce','is_biotech','is_consulting','is_othercategory','has_VC','has_angel','yyyy','mm','month_name','closed_at','first_funding_at','first_funding_at_month'],axis = 1)                

df2

df2.info()

df2['age_last_funding_year'] = df2['age_last_funding_year'].astype(int)
df2['age_first_milestone_year'] = df2['age_first_milestone_year'].astype(int)
df2['age_last_milestone_year'] = df2['age_last_milestone_year'].astype(int)
df2['avg_participants'] = df2['avg_participants'].astype(int)

df2.info()

# pip install sklearn_pandas

X = df2.drop(['status'],axis=1)
Y = df2['status']

Y.shape

#covert numerical format
from sklearn_pandas import DataFrameMapper
from sklearn.preprocessing import LabelEncoder,StandardScaler
from sklearn.ensemble import RandomForestClassifier

df2.to_csv("new.csv")

df2

df2.info()

df2['age_last_funding_year'].unique()

df2['age_first_milestone_year'].unique()

df2['age_last_milestone_year'].unique()

df2['relationships'].unique()

df2['funding_rounds'].unique()

df2['funding_total_usd'].unique()

df2['funding_total_usd'].describe()

df2['milestones'].unique()

df2['is_otherstate'].unique()

df2['avg_participants'].unique()

df2['closed_at_year'].unique()

df2['closed_at_month'].unique()

df2['first_funding_at_year'].unique()

# dm = DataFrameMapper([
#     (["closed_at",LabelEncoder()]),
#     (["first_funding_at",LabelEncoder()])
    
#                       ])

X = df2.drop(['status'],axis=1)
Y = df2['status']

X

Y

ml =[]
# ml.append(['mapper',dm])
ml.append(['sc',StandardScaler()])
ml.append(['RF',RandomForestClassifier()])
ml

from sklearn.pipeline import Pipeline
import warnings
warnings.filterwarnings('ignore')

model = Pipeline(ml)
model.fit(X,Y)

dict(df2.iloc[5,:])

import numpy as np
data ={'age_last_funding_year': np.int64(4),
 'age_first_milestone_year': np.int64(5),
 'age_last_milestone_year': np.int64(5),
 'relationships': np.int64(3),
 'funding_rounds': np.int64(1),
 'funding_total_usd': np.int64(7500000),
 'milestones': np.int64(1),
 'is_otherstate': np.int64(0),
 'has_roundA': np.int64(0),
 'has_roundB': np.int64(1),
 'has_roundC': np.int64(0),
 'has_roundD': np.int64(0),
 'avg_participants': np.int64(3),
 'is_top500': np.int64(1),

 'closed_at_year': np.int64(2024),
 'closed_at_month': np.int64(11),
 'first_funding_at_year': np.int64(2006)}

new_data = pd.DataFrame(data,index=[0])
new_data

model.predict(new_data)

import pickle
with open('final_model.sav',mode='wb')as f:
    pickle.dump(model,f)

df2.describe()

df2

df2.info()
